use identical name for research object store directory as of algorithm object store directory
